+++ packages/invoice-dashboard/src/lib/view-requests.svelte
@@ -41,11 +41,12 @@
   import { CurrencyManager } from "@requestnetwork/currency";
   import { onDestroy, onMount, tick } from "svelte";
   import { formatUnits } from "viem";
-  import { debounce, formatAddress } from "../utils";
+  import { debounce, formatAddress, getEthersSigner } from "../utils";
   import { Drawer, InvoiceView } from "./dashboard";
   import { getPaymentNetworkExtension } from "@requestnetwork/payment-detection";
   import { CipherProviderTypes, CurrencyTypes } from "@requestnetwork/types";
   import { checkStatus } from "@requestnetwork/shared-utils/checkStatus";
+  import { ethers } from "ethers";
 
   export let config: IConfig;
   export let wagmiConfig: WagmiConfig;
@@ -55,7 +56,9 @@
   let cipherProvider: CipherProviderTypes.ICipherProvider | undefined =
     requestNetwork?.getCipherProvider();
 
-  let sliderValueForDecryption = cipherProvider?.isDecryptionEnabled()
+  let sliderValueForDecryption = JSON.parse(
+    localStorage?.getItem("isDecryptionEnabled") ?? "false"
+  )
     ? "on"
     : "off";
 
@@ -102,6 +105,7 @@
   $: {
     if (account?.address) {
       tick().then(() => {
+        enableDecryption();
         getRequests();
       });
     }
@@ -130,6 +134,14 @@
   onDestroy(() => {
     if (typeof unwatchAccount === "function") unwatchAccount();
   });
+
+  $: cipherProvider =
+    requestNetwork?.getCipherProvider() as CipherProviderTypes.ICipherProvider & {
+      getSessionSignatures: (
+        signer: ethers.Signer,
+        walletAddress: `0x${string}`
+      ) => Promise<any>;
+    };
 
   $: {
     signer = account?.address;
@@ -393,15 +405,30 @@
     activeRequest = undefined;
   };
 
-  $: sliderValueForDecryption, getRequests();
-
-  $: {
+  const enableDecryption = async () => {
+    loading = true;
     if (sliderValueForDecryption === "on") {
-      cipherProvider?.enableDecryption(true);
+      try {
+        const signer = await getEthersSigner(wagmiConfig);
+        if (signer && account?.address) {
+          await cipherProvider?.getSessionSignatures(signer, account.address);
+          cipherProvider?.enableDecryption(true);
+          localStorage?.setItem("isDecryptionEnabled", JSON.stringify(true));
+        }
+      } catch (error) {
+        console.error("Failed to enable decryption:", error);
+        toast.error("Failed to enable decryption.");
+        loading = false;
+        return;
+      }
     } else {
       cipherProvider?.enableDecryption(false);
+      localStorage?.setItem("isDecryptionEnabled", JSON.stringify(false));
     }
-  }
+    await getRequests();
+    loading = false;
+  };
+  $: sliderValueForDecryption, enableDecryption();
 </script>
 
 <div