+++ packages/invoice-dashboard/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @requestnetwork/invoice-dashboard
 
+## 0.11.8
+
+### Patch Changes
+
+- Fix enbable decryption
+
+## 0.11.7
+
+### Patch Changes
+
+- Fix InvoiceDashboardProps types
+
+## 0.11.6
+
+### Patch Changes
+
+- Upgrade SDK to 0.53.0
+
 ## 0.11.5
 
 ### Patch Changes